Almost added compile support for windows
Was able to compile mpv at least for windows and a library, i've included the compiled library in res and modified the project file to on windows use the library. This has not been tested and I suspect will not work but in the near future it will. mxe.sh is the script file I used (modified by myself) to compile the build system with necessary libraries, compile mpv and libmpv, and finally (though i haven't successfully done it), compile Baka Mplayer for windows.