Added code for the location dialog
And converted almost all the functions in mainwindow.cpp to signals. But not with the use of lambdas I find it ugly so I'm gonna think of a nicer way to implement the signals and put change it again. Watch it ends up being methods like it was originally... At least the methods will be named better.