More playlist functionality and minor fixes
PlayNext and PlayPrevious methods part of the playlist are now connected to the next/previous buttons. Came up with a way to deal with the playlist--saving the path part of the files to path and displaying to the user only the file-names. After I finish up the playlist (really just programically selecting) I'll have to add mpv url support.
sbin/PKGBUILD is the pkgbuild script I used to create a mxe mingw toolchain on my system. I do not advise anyone else to do it: it took ~2 hours on my Core i7 920 in 16gb of RAM.