Removed some stuff from mpvhandler
Fixed up some things and put in Qt::QueuedConnection for the PlayFile signal. Though this should fix any possible clashing thread issues it still hasn't fixed the issue. I'm going to look into a way to make sure all functions in mpv get called on the same thread, this might be the issue--similarly with the playlist manager.