Added manual and made some strides towards windows compilation
We're almost there--I've successfully compiled libmpv dlls--getting baka mplayer to compile by linking them however is not working for some reason--it tells me the symbols aren't found even though a simple nm command shows that the symbols are there. I think it's as simple as fixing the Makefile but i'll figure it out in due time.