Completed playlist changes and fixed up a few other things
Seek bar wasn't working (probably for a few commits now), I forgot to rename the function; done now.  I made the playlist all work the way it should with it's new model--though it looks much nicer and works much better.. it still holds the same glitch as before. I'm convinced it has something to do with signals & slots which I am in the process of looking into.