Fixed show in folder and cancel problem
Show in Folder didn't work as it was intended because of the mpv->GetFile change; so now we're using the full path from the playlist. (I had fixed this in another branch which i have yet to merge)
Cancel problem fixed, when you cancel open dialog or openUrl dialog, it no longer stops playback.