Added working chapter support
Chapter titles are loaded and they populate the Chapter menu; those menu-items work seeking to the time specified in chapter-list/N/time.
Chapter ticks are loaded by the customslider--considering creating two sliders, turning this custom slider into the seekbar slider and making one for the volume. The volume doesn't need functionality for ticks or tracking so it's minor but unnecessary bloat.