Split up functionality of customslider
Customslider functionality is useful for both the volume slider and the seekbar, but I move the seekbar specific functionality to a new derived class SeekBar.

Furthermore I commented out some aspects of the code which are unnesessary overhead in the customslider that-is, the check for verticle/horizontal. all of our sliders are horizontal.