Using Load() method as an initializer and OnTop support for Linux
We LoadSettings in the load method which is outside of the constructor; this solves various problems that occur because the ui hasn't initialized before the end of the constructor so everything has width/heights of 0 and such. With this things work as they're supposed to.
Using XLib I've completed the AlwaysOnTop implementation for X11; not using Qt api it doesn't flash.