Various Fixes
Fixed index label by disabling it when there are no items and changing the message. Also set the currentRow to the currently playing file when the playlist is changed--this should eliminate any times when there is nothing selected (except when there is nothing in the playlist).
Made it so DimDialog undims when it receives a click--this allowed me to change the functionality and clean it up a bit.