Fixed MpvOverlay when mpv isn't around.
This is done quite simply by creating a QLabel positioned in the same place as the mpv overlay. If mpv is acting--it'll display the stuff for us, if not--the label will.

This requires miniscule overhead--in the future I'll remove even that by "smartly" showing the label only when mpv isn't acting.